Refactor : 권한에 따른 접근 제한 조정
