+++ backend/src/main/java/com/example/coffee/order/controller/AdminOrderController.java
@@ -1,0 +1,38 @@
+package com.example.coffee.order.controller;
+
+import com.example.coffee.order.controller.dto.OrderResponse;
+import com.example.coffee.order.service.OrderService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/admin/orders")
+@RequiredArgsConstructor
+public class AdminOrderController {
+
+    private final OrderService orderService;
+
+    // ✅ 주문 전체 조회
+    @GetMapping
+    @ResponseStatus(HttpStatus.OK)
+    public List<OrderResponse> findAll() {
+        return orderService.findAll();
+    }
+
+    // ✅ 주문 id로 주문 단건 조회
+    @GetMapping("/{id}")
+    @ResponseStatus(HttpStatus.OK)
+    public OrderResponse findById(@PathVariable Long id) {
+        return orderService.findById(id);
+    }
+
+    // ✅ 주문 id로 주문 단건 삭제
+    @DeleteMapping("/{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void delete(@PathVariable Long id) {
+        orderService.delete(id);
+    }
+}
