+++ backend/src/main/java/com/example/coffee/user/controller/UserController.java
@@ -1,6 +1,5 @@
 package com.example.coffee.user.controller;
 
-import com.example.coffee.order.controller.dto.OrderResponse;
 import com.example.coffee.user.controller.dto.CreateUserRequest;
 import com.example.coffee.user.controller.dto.CreateUserResponse;
 import com.example.coffee.user.controller.dto.LoginRequest;
@@ -8,7 +7,6 @@ import com.example.coffee.user.controller.dto.TokenResponse;
 import com.example.coffee.user.controller.dto.RefreshTokenRequest;
 import com.example.coffee.user.service.UserService;
 import lombok.RequiredArgsConstructor;
-import org.apache.catalina.security.SecurityUtil;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
@@ -40,12 +38,6 @@ public class UserController {
         return userService.refreshAccessToken(request.refreshToken());
     }
 
-//    @PostMapping
-//    @ResponseStatus(HttpStatus.CREATED)
-//    public CreateUserResponse createUser(@RequestBody CreateUserRequest request) {
-//        return userService.saveUser(request);
-//    }
-
     // ✅ id로 유저 조회
     @GetMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
@@ -60,24 +52,10 @@ public class UserController {
         return userService.getAllUsers();
     }
 
-//    @PutMapping("/{id}")
-//    @ResponseStatus(HttpStatus.OK)
-//    public CreateUserResponse updateUser(@PathVariable Long id, @RequestBody CreateUserRequest request){
-//        return userService.updateUser(id, request);
-//    }
-
     // ✅ 유저 삭제
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteUser(@PathVariable Long id){
         userService.deleteUser(id);
     }
-
-    // ✅ 해당 유저의 주문 정보 모두 조회
-    // order api와 겹침
-    /*@GetMapping("/{id}/orders")
-    @ResponseStatus(HttpStatus.OK)
-    public List<OrderResponse> getUserOrders(@PathVariable Long id){
-        return userService.getUserOrders(id);
-    }*/
 }
